# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kube API service is service for connection with Kubernetes.",
    "title": "Kube API"
  },
  "paths": {
...
    "/namespaces/{namespace}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Get namespace.",
        "operationId": "GetNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "namespace",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Namespace"
        ],
        "summary": "Update namespace.",
        "operationId": "UpdateNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "namespace updated",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      },
      "delete": {
        "tags": [
          "Namespace"
        ],
        "summary": "Delete namespace.",
        "operationId": "DeleteNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "namespace deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
...
    "/namespaces/{namespace}/pods": {
      "get": {
        "tags": [
          "Pod"
        ],
        "summary": "Get pods list.",
        "operationId": "GetPodList",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "owner",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "pod list",
            "schema": {
              "$ref": "#/definitions/PodsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/pods/{pod}": {
      "get": {
        "tags": [
          "Pod"
        ],
        "summary": "Get pod.",
        "operationId": "GetPod",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "pod",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "pod",
            "schema": {
              "$ref": "#/definitions/PodWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Pod"
        ],
        "summary": "Delete pod.",
        "operationId": "DeletePod",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "pod",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "pod deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/pods/{pod}/log": {
      "get": {
        "tags": [
          "Pod"
        ],
        "summary": "Get pod logs.",
        "operationId": "GetPodLogs",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "$ref": "#/parameters/UpgradeHeader"
          },
          {
            "$ref": "#/parameters/ConnectionHeader"
          },
          {
            "$ref": "#/parameters/SecWebSocketKeyHeader"
          },
          {
            "$ref": "#/parameters/SecWebsocketVersionHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "pod",
            "in": "path",
            "required": true
          },
          {
            "type": "bool",
            "name": "follow",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "tail",
            "in": "query"
          },
          {
            "type": "string",
            "name": "container",
            "in": "query"
          },
          {
            "type": "bool",
            "name": "previous",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "pod logs"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "Container": {
      "description": "Container -- model for container in deployment",
      "type": "object",
      "required": [
        "image",
        "name",
        "limits"
      ],
      "properties": {
        "commands": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Commands"
        },
        "config_maps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerVolume"
          },
          "x-go-name": "ConfigMaps"
        },
        "env": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Env"
          },
          "x-go-name": "Env"
        },
        "image": {
          "type": "string",
          "x-go-name": "Image"
        },
        "limits": {
          "$ref": "#/definitions/Resource"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerPort"
          },
          "x-go-name": "Ports"
        },
        "volume_mounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerVolume"
          },
          "x-go-name": "VolumeMounts"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ContainerPort": {
      "description": "ContainerPort -- model for port in container",
      "type": "object",
      "required": [
        "name",
        "port",
        "protocol"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Port"
        },
        "protocol": {
          "$ref": "#/definitions/Protocol"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ContainerVolume": {
      "description": "ContainerVolume -- volume (or config map) mounted in container",
      "type": "object",
      "required": [
        "name",
        "mount_path"
      ],
      "properties": {
        "mode": {
          "type": "string",
          "x-go-name": "Mode"
        },
        "mount_path": {
          "type": "string",
          "x-go-name": "MountPath"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "pvc_name": {
          "type": "string",
          "x-go-name": "PersistentVolumeClaimName"
        },
        "sub_path": {
          "type": "string",
          "x-go-name": "SubPath"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Env": {
      "description": "Env -- key-value pair of environment variables",
      "type": "object",
      "required": [
        "value",
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
...
    "Pod": {
      "description": "Pod -- model for pod",
      "type": "object",
      "properties": {
        "containers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Container"
          },
          "x-go-name": "Containers"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deploy": {
          "type": "string",
          "x-go-name": "Deploy"
        },
        "hostname": {
          "type": "string",
          "x-go-name": "Hostname"
        },
        "image_pull_secret": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "ImagePullSecret"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "status": {
          "$ref": "#/definitions/PodStatus"
        },
        "total_cpu": {
          "description": "total CPU usage by all containers in this pod",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalCPU"
        },
        "total_memory": {
          "description": "total RAM usage by all containers in this pod",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalMemory"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "PodStatus": {
      "description": "PodStatus -- kubernetes status of pod",
      "type": "object",
      "properties": {
        "phase": {
          "type": "string",
          "x-go-name": "Phase"
        },
        "restart_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RestartCount"
        },
        "start_at": {
          "description": "pod start date in RFC3339 format",
          "type": "string",
          "x-go-name": "StartAt"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "PodsList": {
      "description": "PodsList -- model for pods list",
      "type": "object",
      "properties": {
        "pods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Pod"
          },
          "x-go-name": "Pods"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Protocol": {
      "description": "represents port protocol type (TCP or UDP)",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "Resource": {
      "description": "Resource -- represents namespace CPU and RAM",
      "type": "object",
      "required": [
        "cpu",
        "memory"
      ],
      "properties": {
        "cpu": {
          "description": "CPU in m",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CPU"
        },
        "memory": {
          "description": "RAM in Mi",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Memory"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
  },
  "parameters": {
    "ConnectionHeader": {
      "type": "string",
      "format": "string",
      "name": "Connection",
      "in": "header",
      "required": true
    },
    "SecWebSocketKeyHeader": {
      "type": "string",
      "format": "string",
      "name": "Sec-WebSocket-Key",
      "in": "header",
      "required": true
    },
    "SecWebsocketVersionHeader": {
      "type": "integer",
      "format": "int",
      "name": "Sec-Websocket-Version",
      "in": "header",
      "required": true
    },
    "UpgradeHeader": {
      "type": "string",
      "format": "string",
      "name": "Upgrade",
      "in": "header",
      "required": true
    },
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserNamespaceHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserNamespacesHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}